Set up Jest and Supertest for backend testing and add initial scan endpoint test
