+++ .env.example
@@ -7,7 +7,7 @@ FLASK_API_URL=http://localhost:5000
 PORT=3000
 
 # MongoDB
-MONGO_URI=
+MONGO_URI=mongodb+srv://avishkaa342_db_user:H0WFqIuKJb6axvzu@cluster0.9h6qq3g.mongodb.net/smishing_backend?retryWrites=true&w=majority
 
 # JWT
 JWT_SECRET=to-be-changed-in-production
@@ -16,8 +16,8 @@ JWT_SECRET=to-be-changed-in-production
 EMAIL_HOST=smtp.gmail.com
 EMAIL_PORT=465
 EMAIL_SECURE=true
-EMAIL_USER=
-EMAIL_PASS=
+EMAIL_USER=avishkaa342@gmail.com
+EMAIL_PASS=huyjbiosotnujyly
 
 # OTP Settings
 OTP_LENGTH=6
@@ -36,4 +36,3 @@ feature/ai-chat-api
 
 # ML_Service
 ML_SERVICE_URL=
-
